Added interface for Newline insertion (buggy in libscorepress!). Moved compiler flags from  Makefile  to  configure.ac. Created a  .desktop  file. Added runtime library search path to the executable. (and smaller bugfixes regarding multiple program invocations.)

+++ src/mainwnd.cpp
@@ -243,7 +243,7 @@ void MainWnd::on_zoom_changed()
     if (!engine || view == views.end()) return;
     
     // set zoom parameter
-    engine->get_press_parameters().scale = 50 * zoomScl->get_value();
+    engine->get_press_parameters().scale = static_cast<int>(50.0 * zoomScl->get_value());
     
     // redraw score
     if (view->second->widget.get_realized())
